+++ src/main/java/com/example/locadorabackend/Gerencia/DiretorControler.java
@@ -1,11 +1,15 @@
 package com.example.locadorabackend.Gerencia;
 
 import com.example.locadorabackend.Domain.actor.Actor;
+import com.example.locadorabackend.Domain.classe.Classe;
 import com.example.locadorabackend.Domain.diretor.Diretor;
 import com.example.locadorabackend.Domain.diretor.DiretorRepository;
 import com.example.locadorabackend.Domain.diretor.RequestDiretor;
+import com.example.locadorabackend.Domain.titulo.Titulo;
+import com.example.locadorabackend.Domain.titulo.TituloRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -20,6 +24,9 @@ public class DiretorControler {
 
     @Autowired
     private DiretorRepository diretorRepository;
+
+    @Autowired
+    private TituloRepository tituloRepository;
 
     @GetMapping
     public List<Diretor> getAllDiretor(){
@@ -56,12 +63,19 @@ public class DiretorControler {
     public ResponseEntity<?> excluirDiretor(@PathVariable Long id) {
         Optional<Diretor> optionalDiretor = diretorRepository.findById(id);
         if (optionalDiretor.isPresent()) {
+
+            if (isDiretorAssociadoATitulos(optionalDiretor.get())) {
+                return ResponseEntity.badRequest().body("O Diretor está associada a títulos e não pode ser excluído.");
+            }
             diretorRepository.deleteById(id);
             return ResponseEntity.ok().build();
         } else {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Nao Encontrado no Banco");
         }
     }
 
+    private boolean isDiretorAssociadoATitulos(Diretor diretor){
+        return tituloRepository.existsByDiretor(diretor);
+    }
 
 }
